+++ lib/core/ui/list_view/reorderable_list_view.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:lunasea/core.dart';
 import './flutter_reorderable_list_view.dart';
 
 class LunaReorderableListView extends StatelessWidget {
@@ -28,9 +29,9 @@ class LunaReorderableListView extends StatelessWidget {
         keyboardDismissBehavior: ScrollViewKeyboardDismissBehavior.onDrag,
         children: children,
         padding: padding ??
-            MediaQuery.of(context)
-                .padding
-                .add(EdgeInsets.symmetric(vertical: 8.0)),
+            MediaQuery.of(context).padding.add(EdgeInsets.symmetric(
+                  vertical: LunaUI.MARGIN_CARD.bottom,
+                )),
         physics: physics,
         onReorder: onReorder,
         buildDefaultDragHandles: buildDefaultDragHandles,